Revert "mtd: spi-nor-core: Perform a Soft Reset on boot"
This reverts commit 0be8ab1f166844d53477387dc9a1184161ef44ef.

There are multiple reset commands and it was used one at guess,
see BFPT[dword(16)]. It is preferable to avoid issuing unsupported
commands to a flash. Since there's no config in mainline that actually
uses SPI_FLASH_SOFT_RESET_ON_BOOT, remove it entirely until proper
support is added. One should instead determine the mode in which the
flash device is configured, then to parse SFDP to determine the
cmd_ext_type and then to issue a READID command if flash identification
is really a must. JESD216D-01 proposes an algorithm to try to read the
SFDP signature to determine the mode in which the flash is configured:
'''
try to read the SFDP signature (see 6.1) in 4-4-4 mode, if that fails
try 2-2-2 mode, and if that fails try 1-1-1 mode. For Octal devices,
these typically support SFDP read operation in both 1S-1S-1S mode and
8D-8D-8D mode. If the host controller does not know exactly which
protocol mode is used for SFDP in 8D-8D-8D mode, this information can be
found by reading SFDP in 1S-1S-1S mode first. (To read an unknown device
directly in 8D-8D-8D mode, the host controller may read from address 0,
and count the number of dummy clocks required before the SFDP signature
is received.)
'''
If the flash does not support SFDP at all, one should introduce dedicated
configs for each reset type and issue just the needed reset command.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>