mtd: spi-nor: core: Always reset the flash at remove()
It's a good practice to reset the flash at remove, regardless of its mode
of operation to avoid undefined behaviors. Rest the flash at remove().

It was observed that the hw adjustments between the DT's
spi-tx/rx-bus-width, the SPI controller caps and the flash caps is not
working properly and the flash enters Octal mode even when
spi-tx/rx-bus-width = <1>. This also works as a workaround for the
involuntarily enter in Octal mode.

Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>